Remove workaround for missing signals in VCDStorage.py
We erroneously assumed that signals might not end up in the VCD if they
don't change their value from the default. However, this is not true.
Instead, what's happening is that previous Verilator versions (<4.106)
truncate long signal names and extends them with a hash. This means the
signal is actually there but we can't associate it with a wire in the
netlist.