Add optional `--synthesis-template` arg to parse.py
Even though existing `--synthesis-file` parameter allows one to bring
a custom yosys script, this is not very flexible, because it needs the
source files hardcoded into the script. This is because, this argument
 does not use the yosys patching function `create_yosys_script` in
  `parse.py`.

The new optional `--synthesis-template` allows one to pass
a *template* yosys script that is still patched by `parse.py` so
that source files do not have to be hard coded into the script.
This gives the caller a good flexibility of controlling the synthesis
flow without having to deal with source files.

Ideal use case would be to pass both `--source` and
`--synthesis-template` together.